// autogpt_platform/frontend/src/app/api/__generated__/models/libraryAgentPresetUpdatable.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { LibraryAgentPresetUpdatableInputs } from "./libraryAgentPresetUpdatableInputs";
import type { LibraryAgentPresetUpdatableCredentials } from "./libraryAgentPresetUpdatableCredentials";
import type { LibraryAgentPresetUpdatableName } from "./libraryAgentPresetUpdatableName";
import type { LibraryAgentPresetUpdatableDescription } from "./libraryAgentPresetUpdatableDescription";
import type { LibraryAgentPresetUpdatableIsActive } from "./libraryAgentPresetUpdatableIsActive";

/**
 * Request model used when updating a preset for a library agent.
 */
export interface LibraryAgentPresetUpdatable {
  inputs?: LibraryAgentPresetUpdatableInputs;
  credentials?: LibraryAgentPresetUpdatableCredentials;
  name?: LibraryAgentPresetUpdatableName;
  description?: LibraryAgentPresetUpdatableDescription;
  is_active?: LibraryAgentPresetUpdatableIsActive;
}